requirements.txt
# ============================================================
# AIRFOIL GENERATOR — DEPENDENCIES
# ============================================================
# Install with: pip install -r requirements.txt
# ============================================================

# Core Scientific Computing
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
scipy>=1.7.0
h5py>=3.7.0

# Machine Learning - Classical
scikit-learn>=1.0.0
xgboost>=1.5.0

# Machine Learning - Deep Learning
# PyTorch (install separately if GPU — see instructions)
torch>=1.12.0
torchvision>=0.13.0

# Hyperparameter Optimization
optuna>=3.0.0

# Aerodynamics
# xfoil — may need manual installation on some systems

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Experiment Tracking
tensorboard>=2.10.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
joblib>=1.1.0

# Testing
pytest>=7.0.0

# Notebook
jupyter>=1.0.0
ipywidgets>=8.0.0
nbformat>=5.0.0

# Code Quality
black>=22.0.0
flake8>=5.0.0